backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
asyncpg==0.29.0
pydantic
pydantic-settings
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
langchain==0.2.14
langchain-google-genai==1.0.8
python-multipart==0.0.9
email-validator==2.2.0
chromadb>=0.6.3
